requirements.txt
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
scikit-learn>=1.3.2
xgboost>=2.0.2
entsoe-py>=0.5.10
holidays>=0.38
joblib>=1.3.2
python-dotenv>=1.0.0
notebook>=7.0.6
